+++ apps/explorer/src/comps/Amount.tsx
@@ -2,7 +2,7 @@ import { Link } from '@tanstack/react-router'
 import type { Address } from 'ox'
 import * as Value from 'ox/Value'
 import { maxUint256 } from 'viem'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { useReadContracts } from 'wagmi'
 import { Hooks } from 'wagmi/tempo'
 import { TokenIcon } from '#comps/TokenIcon.tsx'
