+++ apps/explorer/src/comps/TxDecodedTopics.tsx
@@ -2,7 +2,7 @@ import type { AbiEvent } from 'abitype'
 import { useMemo, useState } from 'react'
 import { decodeEventLog, getAbiItem, parseAbiItem } from 'viem'
 import type { Abi, Hex, Log } from 'viem'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { decodeEventLog_guessed, formatAbiValue } from '#lib/domain/contracts'
 import { useCopy } from '#lib/hooks'
 import { useAutoloadAbi, useLookupSignature } from '#lib/queries'
