+++ apps/explorer/src/lib/domain/receipt.ts
@@ -2,7 +2,8 @@ import * as Address from 'ox/Address'
 import { TokenRole } from 'ox/tempo'
 import type { AbiEvent, Log, TransactionReceipt } from 'viem'
 import { parseEventLogs, zeroAddress } from 'viem'
-import { Abis, Addresses } from 'viem/tempo'
+import { Addresses } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { decodeMemoForDisplay } from '#lib/domain/memo'
 import type * as Tip20 from '#lib/domain/tip20'
 import { HexFormatter, PriceFormatter } from '#lib/formatting'
