+++ apps/explorer/src/lib/domain/tip20.ts
@@ -2,7 +2,7 @@ import type { Address } from 'ox'
 import type { Log } from 'viem'
 import { parseEventLogs } from 'viem'
 import { readContract, readContracts } from 'wagmi/actions'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { Actions } from 'wagmi/tempo'
 import type { Config } from 'wagmi'
 import { getWagmiConfig } from '#wagmi.config.ts'
