+++ apps/explorer/src/lib/server/build-tx-only-transactions.ts
@@ -4,7 +4,7 @@ import * as HexUtils from 'ox/Hex'
 import type { Config } from 'wagmi'
 import type { Log, TransactionReceipt } from 'viem'
 import { decodeFunctionData, parseEventLogs } from 'viem'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { Actions } from 'wagmi/tempo'
 import { type KnownEvent, parseKnownEvents } from '#lib/domain/known-events'
 import { isTip20Address, type Metadata } from '#lib/domain/tip20'
