+++ apps/explorer/src/lib/server/fee-amm.ts
@@ -1,7 +1,8 @@
 import { createServerFn } from '@tanstack/react-start'
 import type { Address } from 'ox'
 import * as Value from 'ox/Value'
-import { Abis, Addresses } from 'viem/tempo'
+import { Addresses } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import type { Config } from 'wagmi'
 import { getChainId, readContracts } from 'wagmi/actions'
 import { Actions } from 'wagmi/tempo'
