+++ apps/explorer/src/routes/_layout/block/$id.tsx
@@ -11,7 +11,7 @@ import * as Hex from 'ox/Hex'
 import * as Value from 'ox/Value'
 import * as React from 'react'
 import { decodeFunctionData, isHex, zeroAddress } from 'viem'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { useChains } from 'wagmi'
 import { getBlock } from 'wagmi/actions'
 import * as z from 'zod/mini'
