+++ apps/explorer/src/routes/api/tip20-roles.ts
@@ -3,7 +3,7 @@ import * as Address from 'ox/Address'
 import * as Hash from 'ox/Hash'
 import * as Hex from 'ox/Hex'
 import { formatUnits } from 'viem'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import { getChainId, readContracts } from 'wagmi/actions'
 import { tempoQueryBuilder } from '#lib/server/tempo-queries-provider'
 import { zAddress } from '#lib/zod'
