+++ apps/explorer/src/routes/api/tx/balance-changes/$hash.ts
@@ -3,7 +3,7 @@ import type { Address } from 'ox'
 import type { Log } from 'viem'
 import { getAbiItem, parseEventLogs, zeroAddress } from 'viem'
 import { getTransactionReceipt, readContract } from 'viem/actions'
-import { Abis } from 'viem/tempo'
+import { Abis } from '#lib/abis'
 import * as z from 'zod/mini'
 import { mapWithConcurrency } from '#lib/network.ts'
 import { zHash } from '#lib/zod'
